feat(sound): add preempt flag and pin persisted SoundDefinition size
Replace the implicit high-priority queue path with an explicit `preempt`
flag on SoundDefinition, reusing padding so the struct layout stays
locked at 76 bytes. The size is part of the persisted /cfg/sound_defs.bin
format (v1), so a change would silently corrupt existing files and eat
into the fully-occupied AXI SRAM (ram0) heap; add a static_assert to
guard it.

Also switch BOOT_PING to a plain TONE, mark EMERGENCY as preempt, and
update the services submodule pointer.

+++ src/drivers/sound/sound_synth.cpp
@@ -48,7 +48,7 @@ void Synth::set_unison(uint8_t voices, uint32_t detune_hz) {
   detune_inc = calc_phase_increment(detune_hz);
 }
 
-void Synth::start_tone(uint32_t freq, uint32_t duration_ms, Waveform wf) {
+void Synth::start_tone(uint16_t freq, uint16_t duration_ms, Waveform wf) {
   /* A tone is a single-note sequence without LFO.  Frequencies above 65535 Hz
      (beyond audible) and durations beyond ~65 s are unsupported. */
   notes[0] = {static_cast<uint16_t>(freq), static_cast<uint16_t>(duration_ms), 0, 0};
